Build your first app as a unit converter for temperature (oC to/from oF), distance (km to/from mile), weight (kg to/from lb), and volume (litre to/from gallon).
Your app should have four pairs of textboxes (EditText components), each pair for a conversion type. When the user enters a value in a textbox, the converted value should appear in the corresponding textbox. Note that the conversion should be two-way. For example, if the user enters a value in the textbox for 'mile', the converted value should appear in the textbox for 'km'. But if the user enters a value in the textbox for 'km', the converted value should appear in the textbox for 'mile'.

Your app should also log the text change events. For example, if the user enters a value in the textbox for 'km', you should write in the log that event (the textbox with changed value, the timestamp, and the changed value).